Clean up func sigs
I’ve removed the unused returns from functions with the pattern:

func foo(dst []float, … ) []float

where the returned slice was the same as the source slice, under the
rationale that if people are using this code for tight loops without
reallocation, then if the caller reallocates the return then they will
add inadvertent inefficiency.  Also, they might get the idea that the
input is not modified.  A function with no returned values is obviously
one with only side effects.

I also standardized the names of the input arguments - specifically,
the ones that are modified in place are called “dst”

Where functions had repeated types of inputs, I used the typical go
syntax of func(x, y [type]) instead of func(x [type], y [type])

I also fixed capitalization for the package comment.